\name{dpsdSim-class}
\docType{class}
\alias{dpsdSim-class}

\title{Class "dpsdSim" ~~~ }
\description{Object that results from performing a DPSD simulation with
  dpsdSim(); contains true values and indexes.}

\section{Slots}{
	 \describe{
    \item{\code{cond}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{subj}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{item}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{lag}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{resp}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{muN}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{muS}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{muR}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{alphaN}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{betaN}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{alphaS}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{betaS}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{alphaR}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{betaR}:}{Object of class \code{"numeric"} ~~ }
  }
}

\seealso{dpsdSim()}
\examples{
showClass("dpsdSim")
}
\keyword{classes}
