\name{gammaProbs}
\alias{gammaProbs}
\title{Function gammaProbs}
\description{Returns the probability of making confidence rating
  responses given parameters of gamma signal detection model.}
\usage{
gammaProbs(scale, shape, bounds)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{scale}{Scale of gamma distribution.}
  \item{shape}{Shape of gamma distributuion, usually fixed to 2.0}
  \item{bounds}{Critieria placed on strenght axis.}
}
