\name{rtgamma}
\alias{rtgamma}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Function rtgamma}
\description{Returns random draws from truncated gamma distributuion.}
\usage{
rtgamma(N, shape, scale, a, b)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{N}{Number of samples.}
  \item{shape}{Shape of gamma distribution.}
  \item{scale}{Scale of gamma distributuion.}
  \item{a}{Lower truncation point.}
  \item{b}{Upper truncation point.}
}
