\name{sd10}
\alias{sd10}
\docType{data}
\title{Pratte and Rouder Exp. 2}
\description{
Signal detection dataset that hasn't been published yet. Please do not use.}
\usage{data(sd10)}
\format{
  A data frame with observations on the following 6 variables.
  \describe{
    \item{\code{cond}}{a numeric vector}
    \item{\code{Scond}}{a numeric vector}
    \item{\code{sub}}{a numeric vector}
    \item{\code{item}}{a numeric vector}
    \item{\code{lag}}{a numeric vector}
    \item{\code{resp}}{a numeric vector}
  }
}
\keyword{internal}

