\name{sd13}
\alias{sd13}
\docType{data}
\title{As of yet unpublished dataset}
\description{
Signal detection dataset from Pratte and Rouder Exp. 3. Please do not use.}
\usage{data(sd13)}
\format{
  A data frame with observations on the following 6 variables.
  \describe{
    \item{\code{cond}}{a numeric vector}
    \item{\code{Scond}}{a numeric vector}
    \item{\code{sub}}{a numeric vector}
    \item{\code{item}}{a numeric vector}
    \item{\code{lag}}{a numeric vector}
    \item{\code{resp}}{a numeric vector}
  }
}
\keyword{internal}
