\name{sd6}
\alias{sd6}
\docType{data}
\title{As of yet unpublished dataset}
\description{
Signal detection dataset that hasn't been published yet. Please do not use.}
\usage{data(sd6)}
\format{
  A data frame with 23984 observations on the following 6 variables.
  \describe{
    \item{\code{cond}}{a numeric vector}
    \item{\code{Scond}}{a numeric vector}
    \item{\code{sub}}{a numeric vector}
    \item{\code{item}}{a numeric vector}
    \item{\code{lag}}{a numeric vector}
    \item{\code{resp}}{a numeric vector}
  }
}
\keyword{internal}
