\name{uvsdSim-class}
\docType{class}
\alias{uvsdSim-class}
\title{Class "uvsdSim"}
\description{Class that holds objects from function uvsdSim()}

\section{Slots}{
	 \describe{
    \item{\code{Scond}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{cond}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{subj}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{item}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{lag}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{resp}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{muN}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{muS}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{muS2}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{alphaN}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{betaN}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{alphaS}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{betaS}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{alphaS2}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{betaS2}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{crit}:}{Object of class \code{"matrix"} ~~ }
  }
}
\keyword{classes}
