\name{dpsd-class}
\docType{class}
\alias{dpsd-class}

\title{Class "dpsd" ~~~ }
\description{Holds all information returned from posterior simulations
  of dual-process models}
\section{Slots}{
	 \describe{
    \item{\code{muN}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{alphaN}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{betaN}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{s2alphaN}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{s2betaN}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{thetaN}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{muS}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{alphaS}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{betaS}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{s2alphaS}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{s2betaS}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{thetaS}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{estN}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{estS}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{estR}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{estCrit}:}{Object of class \code{"matrix"} ~~ }
    \item{\code{blockN}:}{Object of class \code{"matrix"} ~~ }
    \item{\code{blockS}:}{Object of class \code{"matrix"} ~~ }
    \item{\code{blockR}:}{Object of class \code{"matrix"} ~~ }
    \item{\code{s.crit}:}{Object of class \code{"array"} ~~ }
    \item{\code{pD}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{DIC}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{M}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{keep}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{b0}:}{Object of class \code{"matrix"} ~~ }
    \item{\code{b0Crit}:}{Object of class \code{"numeric"} ~~ }
  }
}
\keyword{classes}
